Move the tabs and popovers into separate containers
