+++ src/containers/ui/others.svelte
@@ -1,11 +1,5 @@
 <script>
   import Card from 'ui/card.svelte';
-  import Tabs from 'ui/tabs.svelte';
-
-  const tabs = ['Tab 1', 'Tab 2'];
-  const values = ['tab_1', 'tab_2'];
-  let value = values[0];
-  $: console.log(value);
 </script>
 
 <Card id="typo">
@@ -90,40 +84,5 @@
     <hr />
     Divider with text
     <hr data-text="some text" />
-  </div>
-</Card>
-
-<Card id="tabs">
-  <h1>Tabs</h1>
-  <div class="showcase">
-    <Tabs {values} labels={tabs} bind:value={value} name="tabs" />
-  </div>
-</Card>
-
-<!--todo: popover-->
-<Card id="popovers">
-  <h2 class="todo">Popovers</h2>
-  <div class="showcase">
-    <div class="popover-container">
-      <a href="##">simple link</a>
-      <button class="popover btn shadow-1" type="button">
-        <!-- copy -->
-        <svg
-            class="icon mr feather feather-copy"
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            stroke-width="2"
-            stroke-linecap="round"
-            stroke-linejoin="round">
-          <rect x="9" y="9" width="13" height="13" rx="2" ry="2" />
-          <path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1" />
-        </svg>
-        copy
-      </button>
-    </div>
   </div>
 </Card>
